Annotation.inverse_transform()
