+++ openpifpaf/annotation.py
@@ -1,7 +1,10 @@
+import copy
+import math
 import numpy as np
 
 # pylint: disable=import-error
 from .functional import scalar_value_clipped
+from .transforms import utils
 
 
 class Base:
@@ -138,6 +141,52 @@ class Annotation(Base):
         h = np.max(kps[:, 1][m] + joint_scales[m]) - y
         return [x, y, w, h]
 
+    def inverse_transform(self, meta):
+        assert self.fixed_bbox is None
+
+        ann = copy.deepcopy(self)
+
+        # determine rotation parameters
+        angle = -meta['rotation']['angle']
+        rw = meta['rotation']['width']
+        rh = meta['rotation']['height']
+        cangle = math.cos(angle / 180.0 * math.pi)
+        sangle = math.sin(angle / 180.0 * math.pi)
+
+        # rotation
+        if angle != 0.0:
+            xy = ann.data[:, :2]
+            x_old = xy[:, 0].copy() - (rw - 1) / 2
+            y_old = xy[:, 1].copy() - (rh - 1) / 2
+            xy[:, 0] = (rw - 1) / 2 + cangle * x_old + sangle * y_old
+            xy[:, 1] = (rh - 1) / 2 - sangle * x_old + cangle * y_old
+
+        # offset
+        ann.data[:, 0] += meta['offset'][0]
+        ann.data[:, 1] += meta['offset'][1]
+
+        # scale
+        ann.data[:, 0] = ann.data[:, 0] / meta['scale'][0]
+        ann.data[:, 1] = ann.data[:, 1] / meta['scale'][1]
+        ann.joint_scales /= meta['scale'][0]
+
+        assert not np.any(np.isnan(ann.data))
+
+        if meta['hflip']:
+            w = meta['width_height'][0]
+            ann.data[:, 0] = -ann.data[:, 0] + (w - 1)
+            if meta.get('horizontal_swap'):
+                ann.data[:] = meta['horizontal_swap'](ann.data)
+
+        for _, __, c1, c2 in ann.decoding_order:
+            c1[:2] += meta['offset']
+            c2[:2] += meta['offset']
+
+            c1[:2] /= meta['scale']
+            c2[:2] /= meta['scale']
+
+        return ann
+
 
 class AnnotationDet(Base):
     def __init__(self, categories):
@@ -165,6 +214,25 @@ class AnnotationDet(Base):
             'bbox': [round(float(c), 2) for c in self.bbox],
         }
 
+    def inverse_transform(self, meta):
+        ann = copy.deepcopy(self)
+
+        angle = -meta['rotation']['angle']
+        if angle != 0.0:
+            rw = meta['rotation']['width']
+            rh = meta['rotation']['height']
+            ann.bbox = utils.rotate_box(ann.bbox, rw - 1, rh - 1, angle)
+
+        ann.bbox[:2] += meta['offset']
+        ann.bbox[:2] /= meta['scale']
+        ann.bbox[2:] /= meta['scale']
+
+        if meta['hflip']:
+            w = meta['width_height'][0]
+            ann.bbox[0] = -(ann.bbox[0] + ann.bbox[2]) - 1.0 + w
+
+        return ann
+
 
 class AnnotationCrowd(Base):
     def __init__(self, categories):
@@ -188,3 +256,22 @@ class AnnotationCrowd(Base):
             'category': self.category,
             'bbox': [round(float(c), 2) for c in self.bbox],
         }
+
+    def inverse_transform(self, meta):
+        ann = copy.deepcopy(self)
+
+        angle = -meta['rotation']['angle']
+        if angle != 0.0:
+            rw = meta['rotation']['width']
+            rh = meta['rotation']['height']
+            ann.bbox = utils.rotate_box(ann.bbox, rw - 1, rh - 1, angle)
+
+        ann.bbox[:2] += meta['offset']
+        ann.bbox[:2] /= meta['scale']
+        ann.bbox[2:] /= meta['scale']
+
+        if meta['hflip']:
+            w = meta['width_height'][0]
+            ann.bbox[0] = -(ann.bbox[0] + ann.bbox[2]) - 1.0 + w
+
+        return ann
